Transition
It will go to different activity such as it can go from Log In to Sign up and back or to the Dashboard. It can also go from Log In to dashboard to Log in. The dashboard is not complete and we need to figure out how to unable to back button from the phone when on the dashboard.