Submit Event Actvity
Fixed the font and layout of the actual sign up.  Added the submit activity.  From the dashboard, the add event button will take to the submit activity. The buttons are fuctional but the galery.  The camera button work and it will give access to the camera and take the picture. User will have the option to accept the picture or retake the picture. One thing to decided, do we want to display the image before the user decideds to submit.